feat: social 공통 양식으로 수정 1) 기존에 구현했던 코드를 공통 양식에 맞추어 수정함.
