+++ src/main/java/com/skhuedin/skhuedin/controller/BaseController.java
@@ -1,0 +1,13 @@
+package com.skhuedin.skhuedin.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class BaseController {
+
+    @GetMapping("/")
+    public String main() {
+        return "kakaoLogin";
+    }
+}
