+++ src/main/resources/templates/kakaoLogin.html
@@ -5,7 +5,7 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
 </head>
 <body>
-<a href="https://kauth.kakao.com/oauth/authorize?client_id=f4f3a16b864f2046d38669b4f8a2a482&redirect_uri=http://localhost:8080/auth/kakao/callback&response_type=code"><img
+<a href="/auth/kakao"><img
         src="../img/kakao_login_medium_wide.png"></a>
 <a href="/auth/google"><img
         src="../img/kakao_login_medium_wide.png"></a>
